+++ persistent-sqlite/Database/Persist/Sqlite.hs
@@ -4,20 +4,30 @@
 {-# LANGUAGE CPP #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE PatternGuards #-}
 -- | A sqlite backend for persistent.
 --
 -- Note: If you prepend @WAL=off @ to your connection string, it will disable
--- the write-ahead log. For more information, see
--- <https://github.com/yesodweb/persistent/issues/363>.
+-- the write-ahead log. This functionality is now deprecated in favour of using SqliteConnectionInfo.
 module Database.Persist.Sqlite
     ( withSqlitePool
+    , withSqlitePoolInfo
     , withSqliteConn
+    , withSqliteConnInfo
     , createSqlitePool
+    , createSqlitePoolFromInfo
     , module Database.Persist.Sql
     , SqliteConf (..)
+    , SqliteConnectionInfo
+    , mkSqliteConnectionInfo
+    , sqlConnectionStr
+    , walEnabled
+    , fkEnabled
     , runSqlite
+    , runSqliteInfo
     , wrapConnection
+    , wrapConnectionInfo
     , mockMigration
     ) where
 
@@ -26,29 +36,31 @@ import Database.Persist.Sql.Types.Internal (mkPersistBackend)
 
 import qualified Database.Sqlite as Sqlite
 
+import Control.Applicative as A
+import qualified Control.Exception as E
+import Control.Monad (when)
 import Control.Monad.IO.Class (MonadIO (..))
 import Control.Monad.Logger (NoLoggingT, runNoLoggingT, MonadLogger)
-import Data.IORef
-import qualified Data.Map as Map
 import Control.Monad.Trans.Control (control)
+import Control.Monad.Trans.Control (MonadBaseControl)
+import Control.Monad.Trans.Reader (ReaderT, runReaderT)
+import Control.Monad.Trans.Resource (ResourceT, runResourceT)
+import Control.Monad.Trans.Writer (runWriterT)
 import Data.Acquire (Acquire, mkAcquire, with)
-import qualified Control.Exception as E
-import Data.Text (Text)
 import Data.Aeson
 import Data.Aeson.Types (modifyFailure)
-import qualified Data.Text as T
-import qualified Data.Text.IO as TIO
 import Data.Conduit
 import qualified Data.Conduit.List as CL
-import Control.Applicative
+import qualified Data.HashMap.Lazy as HashMap
 import Data.Int (Int64)
+import Data.IORef
+import qualified Data.Map as Map
 import Data.Monoid ((<>))
 import Data.Pool (Pool)
-import Control.Monad.Trans.Control (MonadBaseControl)
-import Control.Monad.Trans.Resource (ResourceT, runResourceT)
-import Control.Monad (when)
-import Control.Monad.Trans.Reader (ReaderT, runReaderT)
-import Control.Monad.Trans.Writer (runWriterT)
+import Data.Text (Text)
+import qualified Data.Text as T
+import qualified Data.Text.IO as TIO
+import Lens.Micro.TH (makeLenses)
 
 -- | Create a pool of SQLite connections.
 --
@@ -57,7 +69,18 @@ import Control.Monad.Trans.Writer (runWriterT)
 -- Instead, use 'withSqliteConn'.
 createSqlitePool :: (MonadIO m, MonadLogger m, MonadBaseControl IO m, IsSqlBackend backend)
                  => Text -> Int -> m (Pool backend)
-createSqlitePool s = createSqlPool $ open' s
+createSqlitePool = createSqlitePoolFromInfo . conStringToInfo
+
+-- | Create a pool of SQLite connections.
+--
+-- Note that this should not be used with the @:memory:@ connection string, as
+-- the pool will regularly remove connections, destroying your database.
+-- Instead, use 'withSqliteConn'.
+--
+-- @since 2.6.2
+createSqlitePoolFromInfo :: (MonadIO m, MonadLogger m, MonadBaseControl IO m, IsSqlBackend backend)
+                         => SqliteConnectionInfo -> Int -> m (Pool backend)
+createSqlitePoolFromInfo connInfo = createSqlPool $ open' connInfo
 
 -- | Run the given action with a connection pool.
 --
@@ -66,43 +89,64 @@ withSqlitePool :: (MonadBaseControl IO m, MonadIO m, MonadLogger m, IsSqlBackend
                => Text
                -> Int -- ^ number of connections to open
                -> (Pool backend -> m a) -> m a
-withSqlitePool s = withSqlPool $ open' s
+withSqlitePool connInfo = withSqlPool . open' $ conStringToInfo connInfo
+
+-- | Run the given action with a connection pool.
+--
+-- Like 'createSqlitePool', this should not be used with @:memory:@.
+--
+-- @since 2.6.2
+withSqlitePoolInfo :: (MonadBaseControl IO m, MonadIO m, MonadLogger m, IsSqlBackend backend)
+               => SqliteConnectionInfo
+               -> Int -- ^ number of connections to open
+               -> (Pool backend -> m a) -> m a
+withSqlitePoolInfo connInfo = withSqlPool $ open' connInfo
 
 withSqliteConn :: (MonadBaseControl IO m, MonadIO m, MonadLogger m, IsSqlBackend backend)
                => Text -> (backend -> m a) -> m a
-withSqliteConn = withSqlConn . open'
+withSqliteConn = withSqliteConnInfo . conStringToInfo
 
-open' :: (IsSqlBackend backend) => Text -> LogFunc -> IO backend
-open' connStr logFunc = do
-    let (connStr', enableWal) = case () of
-          ()
-            | Just cs <- T.stripPrefix "WAL=on "  connStr -> (cs, True)
-            | Just cs <- T.stripPrefix "WAL=off " connStr -> (cs, False)
-            | otherwise                                   -> (connStr, True)
+-- | @since 2.6.2
+withSqliteConnInfo :: (MonadBaseControl IO m, MonadIO m, MonadLogger m, IsSqlBackend backend)
+                   => SqliteConnectionInfo -> (backend -> m a) -> m a
+withSqliteConnInfo = withSqlConn . open'
 
-    conn <- Sqlite.open connStr'
-    wrapConnectionWal enableWal conn logFunc
+open' :: (IsSqlBackend backend) => SqliteConnectionInfo -> LogFunc -> IO backend
+open' connInfo logFunc = do
+    conn <- Sqlite.open $ _sqlConnectionStr connInfo
+    wrapConnectionInfo connInfo conn logFunc `E.onException` Sqlite.close conn
 
 -- | Wrap up a raw 'Sqlite.Connection' as a Persistent SQL 'Connection'.
 --
 -- Since 1.1.5
 wrapConnection :: (IsSqlBackend backend) => Sqlite.Connection -> LogFunc -> IO backend
-wrapConnection = wrapConnectionWal True
+wrapConnection = wrapConnectionInfo (mkSqliteConnectionInfo "")
 
--- | Allow control of WAL settings when wrapping
-wrapConnectionWal :: (IsSqlBackend backend)
-                  => Bool -- ^ enable WAL?
+-- | Wrap up a raw 'Sqlite.Connection' as a Persistent SQL
+-- 'Connection', allowing full control over WAL and FK constraints.
+--
+-- @since 2.6.2
+wrapConnectionInfo :: (IsSqlBackend backend)
+                  => SqliteConnectionInfo
                   -> Sqlite.Connection
                   -> LogFunc
                   -> IO backend
-wrapConnectionWal enableWal conn logFunc = do
-    when enableWal $ do
+wrapConnectionInfo connInfo conn logFunc = do
+    when (_walEnabled connInfo) $ do
         -- Turn on the write-ahead log
         -- https://github.com/yesodweb/persistent/issues/363
         turnOnWal <- Sqlite.prepare conn "PRAGMA journal_mode=WAL;"
         _ <- Sqlite.step turnOnWal
         Sqlite.reset conn turnOnWal
         Sqlite.finalize turnOnWal
+
+    when (_fkEnabled connInfo) $ do
+        -- Turn on foreign key constraints
+        -- https://github.com/yesodweb/persistent/issues/646
+        turnOnFK <- Sqlite.prepare conn "PRAGMA foreign_keys = on;"
+        _ <- Sqlite.step turnOnFK
+        Sqlite.reset conn turnOnFK
+        Sqlite.finalize turnOnFK
 
     smap <- newIORef $ Map.empty
     return . mkPersistBackend $ SqlBackend
@@ -121,6 +165,7 @@ wrapConnectionWal enableWal conn logFunc = do
         , connRDBMS = "sqlite"
         , connLimitOffset = decorateSQLWithLimitOffset "LIMIT -1"
         , connLogFunc = logFunc
+        , connMaxParams = Just 999
         }
   where
     helper t getter = do
@@ -142,6 +187,20 @@ runSqlite connstr = runResourceT
                   . runNoLoggingT
                   . withSqliteConn connstr
                   . runSqlConn
+
+-- | A convenience helper which creates a new database connection and runs the
+-- given block, handling @MonadResource@ and @MonadLogger@ requirements. Note
+-- that all log messages are discarded.
+--
+-- @since 2.6.2
+runSqliteInfo :: (MonadBaseControl IO m, MonadIO m, IsSqlBackend backend)
+              => SqliteConnectionInfo
+              -> ReaderT backend (NoLoggingT (ResourceT m)) a -- ^ database action
+              -> m a
+runSqliteInfo conInfo = runResourceT
+                      . runNoLoggingT
+                      . withSqliteConnInfo conInfo
+                      . runSqlConn
 
 prepare' :: Sqlite.Connection -> Text -> IO Statement
 prepare' conn sql = do
@@ -285,6 +344,8 @@ mockMigration mig = do
                    , connRDBMS = "sqlite"
                    , connLimitOffset = decorateSQLWithLimitOffset "LIMIT -1"
                    , connLogFunc = undefined
+                   , connUpsertSql = undefined
+                   , connMaxParams = Just 999
                    }
       result = runReaderT . runWriterT . runWriterT $ mig
   resp <- result sqlbackend
@@ -426,21 +487,31 @@ escape (DBName s) =
     go '"' = "\"\""
     go c = T.singleton c
 
--- | Information required to connect to a sqlite database
+-- | Information required to setup a connection pool.
 data SqliteConf = SqliteConf
     { sqlDatabase :: Text
+    , sqlPoolSize :: Int
+    }
+    | SqliteConfInfo
+    { sqlConnInfo :: SqliteConnectionInfo
     , sqlPoolSize :: Int
     } deriving Show
 
 instance FromJSON SqliteConf where
-    parseJSON v = modifyFailure ("Persistent: error loading Sqlite conf: " ++) $
-      flip (withObject "SqliteConf") v $ \o -> SqliteConf
-        <$> o .: "database"
-        <*> o .: "poolsize"
+    parseJSON v = modifyFailure ("Persistent: error loading Sqlite conf: " ++) $ flip (withObject "SqliteConf") v parser where
+        parser o = if HashMap.member "database" o
+                      then SqliteConf
+                            A.<$> o .: "database"
+                            A.<*> o .: "poolsize"
+                      else SqliteConfInfo
+                            A.<$> o .: "connInfo"
+                            A.<*> o .: "poolsize"
+
 instance PersistConfig SqliteConf where
     type PersistConfigBackend SqliteConf = SqlPersistT
     type PersistConfigPool SqliteConf = ConnectionPool
-    createPoolConfig (SqliteConf cs size) = runNoLoggingT $ createSqlitePool cs size -- FIXME
+    createPoolConfig (SqliteConf cs size) = runNoLoggingT $ createSqlitePoolFromInfo (conStringToInfo cs) size -- FIXME
+    createPoolConfig (SqliteConfInfo info size) = runNoLoggingT $ createSqlitePoolFromInfo info size -- FIXME
     runPool _ = runSqlPool
     loadConfig = parseJSON
 
@@ -452,3 +523,37 @@ finally a sequel = control $ \runInIO ->
                      E.finally (runInIO a)
                                (runInIO sequel)
 {-# INLINABLE finally #-}
+-- | Creates a SqliteConnectionInfo from a connection string, with the
+-- default settings.
+--
+-- @since 2.6.2
+mkSqliteConnectionInfo :: Text -> SqliteConnectionInfo
+mkSqliteConnectionInfo fp = SqliteConnectionInfo fp True True
+
+-- | Parses connection options from a connection string. Used only to provide deprecated API.
+conStringToInfo :: Text -> SqliteConnectionInfo
+conStringToInfo connStr = SqliteConnectionInfo connStr' enableWal True where
+    (connStr', enableWal) = case () of
+        ()
+            | Just cs <- T.stripPrefix "WAL=on "  connStr -> (cs, True)
+            | Just cs <- T.stripPrefix "WAL=off " connStr -> (cs, False)
+            | otherwise                                   -> (connStr, True)
+
+-- | Information required to connect to a sqlite database. We export
+-- lenses instead of fields to avoid being limited to the current
+-- implementation.
+--
+-- @since 2.6.2
+data SqliteConnectionInfo = SqliteConnectionInfo
+    { _sqlConnectionStr :: Text -- ^ connection string for the database. Use @:memory:@ for an in-memory database.
+    , _walEnabled :: Bool -- ^ if the write-ahead log is enabled - see https://github.com/yesodweb/persistent/issues/363.
+    , _fkEnabled :: Bool -- ^ if foreign-key constraints are enabled.
+    } deriving Show
+makeLenses ''SqliteConnectionInfo
+
+instance FromJSON SqliteConnectionInfo where
+    parseJSON v = modifyFailure ("Persistent: error loading SqliteConnectionInfo: " ++) $
+      flip (withObject "SqliteConnectionInfo") v $ \o -> SqliteConnectionInfo
+        <$> o .: "connectionString"
+        <*> o .: "walEnabled"
+        <*> o .: "fkEnabled"