+++ persistent-test/test/main.hs
@@ -1,54 +1,52 @@
 {-# LANGUAGE CPP #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 
-import qualified PersistentTest
-import qualified RenameTest
-import qualified DataTypeTest
-import qualified EmptyEntityTest
-import qualified HtmlTest
-import qualified EmbedTest
-import qualified EmbedOrderTest
-import qualified LargeNumberTest
-import qualified MaxLenTest
-import qualified Recursive
-import qualified SumTypeTest
-import qualified UniqueTest
-import qualified MigrationTest
-import qualified MigrationOnlyTest
-import qualified PersistUniqueTest
 import qualified CompositeTest
-import qualified PrimaryTest
 import qualified CustomPersistFieldTest
 import qualified CustomPrimaryKeyReferenceTest
-import Test.Hspec (hspec)
-import Test.Hspec.Runner
+import qualified DataTypeTest
+import qualified EmbedOrderTest
+import qualified EmbedTest
+import qualified EmptyEntityTest
+import qualified HtmlTest
 import Init
-import System.Exit
-import Control.Monad (unless, when)
-import Filesystem (isFile, removeFile)
-import Filesystem.Path.CurrentOS (fromText)
-import Control.Monad.Trans.Resource (runResourceT)
+import qualified LargeNumberTest
+import qualified MaxLenTest
+import qualified MigrationOnlyTest
+import qualified PersistentTest
+import qualified PersistUniqueTest
+import qualified PrimaryTest
+import qualified Recursive
+import qualified RenameTest
+import qualified SumTypeTest
+import qualified UniqueTest
+
+#ifndef WITH_NOSQL
+#  ifdef WITH_SQLITE
 import Control.Exception (handle, IOException)
+import Filesystem (removeFile)
+import Filesystem.Path.CurrentOS (fromText)
+import qualified MigrationTest
+#  endif
+#endif
 
 
 #ifdef WITH_NOSQL
 #else
-import Database.Persist.Sql (printMigration, runMigrationUnsafe)
 
+setup :: MonadIO m => Migration -> ReaderT SqlBackend m ()
 setup migration = do
   printMigration migration
   runMigrationUnsafe migration
 #endif
 
-toExitCode :: Bool -> ExitCode
-toExitCode True  = ExitSuccess
-toExitCode False = ExitFailure 1
-
 main :: IO ()
 main = do
 #ifndef WITH_NOSQL
+#  ifdef WITH_SQLITE
   handle (\(_ :: IOException) -> return ())
-    $ removeFile $ fromText sqlite_database
+    $ removeFile $ fromText sqlite_database_file
+#  endif
 
   runConn $ do
     mapM_ setup
@@ -61,7 +59,9 @@ main = do
       , MaxLenTest.maxlenMigrate
       , Recursive.recursiveMigrate
       , CompositeTest.compositeMigrate
+#  ifdef WITH_SQLITE
       , MigrationTest.migrationMigrate
+#  endif
       , PersistUniqueTest.migration
       , RenameTest.migration
       , CustomPersistFieldTest.customFieldMigrate
@@ -93,6 +93,6 @@ main = do
     CustomPersistFieldTest.specs
     CustomPrimaryKeyReferenceTest.specs
 
-#ifndef WITH_NOSQL
+#ifdef WITH_SQLITE
     MigrationTest.specs
 #endif